commands: move sparse code to sparse.rs
